+++ src/components/dialog/dialog.js
@@ -124,7 +124,7 @@ class Dialog extends Modal {
    * @method onOpening
    * @return {Void}
    */
-  get onOpening() {
+  onOpening() {
     this.centerDialog();
     this.focusDialog();
     this.window().addEventListener('resize', this.centerDialog);
@@ -138,7 +138,7 @@ class Dialog extends Modal {
    * @method onClosing
    * @return {Void}
    */
-  get onClosing() {
+  onClosing() {
     this.window().removeEventListener('resize', this.centerDialog);
   }