+++ CHANGELOG.md
@@ -151,6 +151,10 @@ A new helper object is available in `utils/helpers/text`. Currently it only cont
   * has a new prop `onSlideChange`, which is an action to be called on slide change. It will receive the slide index and the transition direction as params.
   * animation between slides enhanced.
 * `Dialog` is now using the `Heading` component to render its title and subtitle.
+* `DialogFullScreen`:
+  * the content is now scrollable.
+  * is now using the `Heading` component to render its title.
+  * the title can be passed as a child instead of using the title prop.
 * `Form` now has default `SaveButton` and `CancelButton` functional stateless components. The former can be overridden with a new prop of `customSaveButton`.
 * `Heading` backLink prop can now be a string or a function.
 * `InlineInputs` children are now wrapped by Columns by the component.